+++ collagraph/renderers/pyside/widget.py
@@ -5,7 +5,7 @@ from PySide6.QtWidgets import (
     QSpacerItem,
 )
 
-from . import camel_case, name_to_type
+from . import attr_name_to_method_name, call_method
 
 
 DIRECTIONS = {
@@ -113,27 +113,3 @@ def set_attribute(self, attr, value):
         return
 
     call_method(method, value)
-
-
-def attr_name_to_method_name(name, setter=False):
-    sep = "-"
-    if "_" in name:
-        sep = "_"
-
-    prefix = f"set{sep}" if setter else ""
-    return camel_case(f"{prefix}{name}", sep)
-
-
-def call_method(method, args):
-    if isinstance(args, str):
-        try:
-            args = name_to_type(args)
-        except TypeError:
-            pass
-        method(args)
-    else:
-        try:
-            method(args)
-        except TypeError:
-            # TODO: Maybe also call name_to_type on all values?
-            method(*args)
